requirements.txt
python-telegram-bot==21.5
APScheduler==3.10.4
Flask==3.0.3
pytz==2024.1
